Replaced GalleryController code with GalleryTest code. Replaced GalleryViews.
