--- Gallery.Web/Views/GalleryTest/ViewGallery.cshtml
@@ -1,34 +1,0 @@
-﻿@model Gallery.Web.gallery
-
-@{
-    ViewBag.Title = "ViewGallery";
-}
-@if (Model != null)
-{
-    <h2>@Model.GalleryName</h2>
-    <h3>Averate Rating: @ViewBag.galleryRating</h3>
-    <p>Likes: @ViewBag.galleryUpVotes</p>
-    <p>Dislikes: @ViewBag.galleryDownVotes</p>
-    <br/>
-    <br/>
-
-    if (Model != null)
-    {
-        foreach (var item in Model.gallery_images)
-        {
-            if (item.image.IsHidden != 1)
-            {
-                var imgLink = item.image.site_locations.Location + item.image.Filename;
-                var img = item.image.site_locations.Location + item.image.Filename + "?w=200&h=300&mode=crop";
-
-                <a href="@Url.Action("ViewImage", "Images", new {Id = item.ImageID}, null)">
-                    <img src="@img"/>
-                </a>
-            }
-        }
-    }
-}
-else
-{
-    <h2>Nothing to see!</h2>
-}
